clean: move dict implementations out of headers
class with Q_OBJECT can be defined .cpp file by including .moc
explicitly
