+++ src/dict/lingualibre.cc
@@ -7,6 +7,7 @@
 
 #include <string>
 #include <utility>
+#include <QNetworkReply>
 
 namespace Lingua {
 
@@ -14,6 +15,47 @@ using namespace Dictionary;
 
 namespace {
 
+class LinguaArticleRequest: public Dictionary::DataRequest
+{
+  Q_OBJECT
+
+  struct NetReply
+  {
+    sptr< QNetworkReply > reply;
+    string word;
+    bool finished;
+
+    NetReply( sptr< QNetworkReply > const & reply_, string const & word_ ):
+      reply( reply_ ),
+      word( word_ ),
+      finished( false )
+    {
+    }
+  };
+
+  typedef std::list< NetReply > NetReplies;
+  NetReplies netReplies;
+  QString languageCode, langWikipediaID;
+  string dictionaryId;
+
+public:
+
+  LinguaArticleRequest( wstring const & word,
+                        vector< wstring > const & alts,
+                        QString const & languageCode_,
+                        QString const & langWikipediaID_,
+                        string const & dictionaryId_,
+                        QNetworkAccessManager & mgr );
+
+  virtual void cancel();
+
+private:
+
+  void addQuery( QNetworkAccessManager & mgr, wstring const & word );
+
+private slots:
+  virtual void requestFinished( QNetworkReply * );
+};
 
 class LinguaDictionary: public Dictionary::Class
 {
@@ -338,5 +380,5 @@ void LinguaArticleRequest::requestFinished( QNetworkReply * r )
   }
 }
 
-
+#include "lingualibre.moc"
 } // end namespace Lingua
